Refactor sync logic
