+++ packages/scripts/src/helpers/sync/index.ts
@@ -1,15 +1,10 @@
-import path from 'path';
 import { updateReadme, ensureOverview } from '../docs/overview';
 import { listPackages, updatePkgJSON } from '../packages';
 import { updateSidebarJSON } from '../docs/sidebar';
-import { formatList, writePkgHeader } from '../misc';
-import { getChangedFiles } from '../scripts';
 import { PackageInfo } from '../interfaces';
-import signale from '../signale';
-
-export type SyncOptions = {
-    verify?: boolean;
-};
+import { SyncOptions } from './interfaces';
+import { verify, getFiles } from './utils';
+import { writePkgHeader } from '../misc';
 
 export async function syncAll(options: SyncOptions = {}) {
     for (const pkgInfo of listPackages()) {
@@ -36,21 +31,4 @@ export async function syncPackages(pkgInfos: PackageInfo[], options: SyncOptions
     );
 
     await verify(files, options);
-}
-
-export async function verify(files: string[], options: SyncOptions) {
-    if (!options.verify) return;
-
-    const changed = await getChangedFiles(...files);
-    if (changed.length) {
-        signale.error(`Files have either changes or are out-of-sync, run 'yarn sync' and push up the changes:${formatList(changed)}`);
-        process.exit(1);
-    }
-}
-
-function getFiles(pkgInfo?: PackageInfo): string[] {
-    if (pkgInfo) {
-        return [path.join('packages', pkgInfo.folderName), path.join('docs/packages', pkgInfo.folderName)];
-    }
-    return ['packages', 'docs', 'website'];
 }
