+++ packages/scripts/src/helpers/sync/interfaces.ts
@@ -1,0 +1,3 @@
+export type SyncOptions = {
+    verify?: boolean;
+};
